+++ terminator/trunk/src/terminator/view/JTerminalPane.java
@@ -653,7 +653,7 @@ public class JTerminalPane extends JPanel {
 			return true;
 		}
 
-		final int directChildPid = ptyProcess.getPid();
+		final long directChildPid = ptyProcess.getPid();
 		final String processesUsingTty = ptyProcess.listProcessesUsingTty();
 
 		if (processesUsingTty.length() == 0) {
